L7 fixes + L8: wallet + GUI send form
L7 fixes:
- thiserror named-args: replace {0} with {idx} where struct variants
  used both positional and named placeholders.
- drop redundant tx.amount as u128 cast.

L8 (Wallet + GUI):
- New crates/gui/src/wallet.rs: Wallet generates an Ed25519 keypair,
  derives the bech32m address, persists JSON to OS-appropriate config
  dir (Phase A plaintext; Phase B encrypts).
- crates/gui/src/rpc_client.rs: get_account + submit_tx wrappers
  separate from miner.rs — wallet path stays clean.
- main.rs loads/creates the wallet on startup, populates the
  Wallet tab (address, balance, nonce), polls balance every 3s,
  wires the Send button to: parse pyg1 recipient + sat amount + fee,
  fetch nonce via RPC, build TxBody, sign signing_hash, package
  Witness (Inline pubkey first time, Known after), CBOR-encode both,
  hex-encode, submit_tx via RPC.
- app.slint: real wallet display + send form (to / amount / fee /
  status), all bound to in-out properties.

+++ crates/state/src/apply.rs
@@ -30,13 +30,13 @@ pub enum ApplyError {
     PubKeyMismatch(usize),
     #[error("tx[{0}]: bad signature ({1})")]
     BadSignature(usize, String),
-    #[error("tx[{0}]: nonce mismatch — got {got}, expected {expected}")]
+    #[error("tx[{idx}]: nonce mismatch — got {got}, expected {expected}")]
     NonceMismatch {
         idx: usize,
         got: u64,
         expected: u64,
     },
-    #[error("tx[{0}]: insufficient balance ({balance} < amount+fee {required})")]
+    #[error("tx[{idx}]: insufficient balance ({balance} < amount+fee {required})")]
     InsufficientBalance {
         idx: usize,
         balance: u128,
@@ -46,7 +46,7 @@ pub enum ApplyError {
     AmountOverflow(usize),
     #[error("tx[{0}]: TxCall variant not supported in Phase A")]
     UnsupportedCall(usize),
-    #[error("tx[{0}]: pubkey algo {algo} disagrees with sig_algo {sig_algo}")]
+    #[error("tx[{idx}]: pubkey algo {algo} disagrees with sig_algo {sig_algo}")]
     AlgoMismatch {
         idx: usize,
         algo: u8,
@@ -109,7 +109,8 @@ pub fn apply_block(
                 expected: tx.expected_nonce,
             });
         }
-        let required = (tx.amount as u128)
+        let required = tx
+            .amount
             .checked_add(tx.fee_sat as u128)
             .ok_or(ApplyError::AmountOverflow(tx.idx))?;
         if from.balance < required {
